Rename example service proto files
