+++ src/assets/thirdPartyServices/ExampleService/example_service_pb.js
